knn 适配 google test 框架
