+++ knn/knn_utils.cpp
@@ -1,18 +1,136 @@
 #include "knn_utils.h"
 
+#include <algorithm>
+#include <chrono>
+#include <cmath>
+#include <fstream>
+#include <iostream>
+#include <random>
+#include <sstream>
+#include <stdexcept>
+
 using json = nlohmann::json;
+using namespace poseidon;
 using namespace poseidon::util;
 
 namespace KNN
 {
+namespace
+{
+constexpr int kDataNums = 100;
+constexpr int kDimension = 10;
+constexpr int kScaleBits = 32;
+constexpr int kNumThreads = 8;
+constexpr int kPredictionCount = 10;
+constexpr int kTopKThreshold = 100;
+constexpr double kInputScaleDivisor = 40.0;
+constexpr double kComparisonThreshold = 10.5;
+constexpr double kFinalScaleFactor = 0.014;
+constexpr double kMarginMaskMin = 0.95;
+constexpr double kMarginMaskMax = 1.05;
+const std::filesystem::path kCurrentPath(__FILE__);
+const std::vector<int> kRotationSteps = {100, 200, 300, 600, 1200, 2500, 5000};
 
-int data_nums = 100;
-int dimension = 10;
-int NUM = 128;
-int num_threads = 8;
+PolynomialVector build_polynomial_vector(const ComplexVec &coeffs,
+                                         std::vector<std::vector<int>> &slots_index,
+                                         int degree)
+{
+    Polynomial poly(coeffs, 0, 0, degree, Monomial);
+    poly.lead() = true;
+    return PolynomialVector(std::vector<Polynomial>{poly}, slots_index);
+}
 
-void read_jsonl_query(const std::string &file,
-                      std::vector<std::vector<std::complex<double>>> &query)
+Ciphertext sum_distance_columns(const std::shared_ptr<EvaluatorCkksBase> &ckks_eva,
+                                const std::vector<Ciphertext> &ciph_data,
+                                size_t start_index)
+{
+    Ciphertext distance = ciph_data[start_index];
+    for (size_t i = 1; i < kDimension; ++i)
+    {
+        ckks_eva->add(distance, ciph_data[start_index + i], distance);
+    }
+    return distance;
+}
+
+std::vector<double> generate_margin_mask(int n)
+{
+    std::vector<double> result;
+    if (n <= 0)
+    {
+        return result;
+    }
+
+    result.reserve(n);
+    unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
+    std::mt19937 generator(seed);
+    std::uniform_real_distribution<double> distribution(kMarginMaskMin, kMarginMaskMax);
+
+    for (int i = 0; i < n; ++i)
+    {
+        result.push_back(distribution(generator));
+    }
+
+    return result;
+}
+} // namespace
+
+RuntimeOptions make_default_options()
+{
+    RuntimeOptions options;
+    options.dataset_file = (kCurrentPath.parent_path() / "dataset" / "train.jsonl").string();
+    options.predictions_file = (kCurrentPath.parent_path() / "predictions.jsonl").string();
+    return options;
+}
+
+RuntimeOptions parse_options(int argc, char *argv[])
+{
+    RuntimeOptions options = make_default_options();
+    for (int i = 1; i < argc; ++i)
+    {
+        const std::string arg = argv[i];
+        if (arg == "--log-degree" && i + 1 < argc)
+        {
+            options.log_degree = static_cast<uint32_t>(std::stoul(argv[++i]));
+        }
+        else if (arg == "--dataset" && i + 1 < argc)
+        {
+            options.dataset_file = argv[++i];
+        }
+        else if (arg == "--predictions" && i + 1 < argc)
+        {
+            options.predictions_file = argv[++i];
+        }
+        else
+        {
+            std::cerr << "Usage: " << argv[0]
+                      << " [--log-degree 15|16] [--dataset path] [--predictions path]" << std::endl;
+            throw std::invalid_argument("invalid command-line arguments");
+        }
+    }
+
+    if (options.log_degree != 15 && options.log_degree != 16)
+    {
+        throw std::invalid_argument("Only --log-degree 15 (32768) and 16 (65536) are supported.");
+    }
+
+    return options;
+}
+
+std::vector<int> decode_predictions(const ComplexVec &decoded)
+{
+    std::vector<int> result;
+    result.reserve(kDataNums);
+    for (int i = 0; i < kDataNums; ++i)
+    {
+        if (std::round(decoded[i].real()) == 1)
+        {
+            result.push_back(i + 1);
+        }
+    }
+    return result;
+}
+
+void read_jsonl_query(const std::string &file, ComplexMatrix &query, size_t slot_count)
 {
     std::ifstream infile(file);
     if (!infile.is_open())
@@ -20,26 +138,35 @@ void read_jsonl_query(const std::string &file,
         POSEIDON_THROW(config_error, "cannot open file ：" + file);
     }
 
-    // 读取整个文件内容
     std::stringstream buffer;
     buffer << infile.rdbuf();
     json j_all = json::parse(buffer.str());
-    infile.close();
 
-    query.clear();
-    // 获取 query 数组
-    std::vector<double> real_vec = j_all["query"].get<std::vector<double>>();
-
-    // 生成 10 行，每行复制对应实数 100 次，转为复数
-    for (int i = 0; i < 10; ++i)
+    if (!j_all.contains("query") || !j_all["query"].is_array())
     {
-        std::vector<std::complex<double>> row(10000, std::complex<double>(real_vec[i] / 40, 0.0));
-        query.push_back(row);
+        throw std::runtime_error("Invalid JSON: 'query' field missing or not an array");
+    }
+
+    const std::vector<double> real_vec = j_all["query"].get<std::vector<double>>();
+    if (real_vec.size() != static_cast<size_t>(kDimension))
+    {
+        throw std::runtime_error("Invalid JSON: 'query' size does not match kDimension");
+    }
+
+    if (slot_count < static_cast<size_t>(kDataNums * kDataNums))
+    {
+        throw std::runtime_error("slot_count is smaller than the packed query size");
+    }
+
+    query.assign(kDimension, ComplexVec(slot_count, {0.0, 0.0}));
+    for (int i = 0; i < kDimension; ++i)
+    {
+        const auto value = Complex(real_vec.at(i) / kInputScaleDivisor, 0.0);
+        std::fill_n(query[i].begin(), kDataNums * kDataNums, value);
     }
 }
 
-void read_jsonl_data(const std::string &file,
-                     std::vector<std::vector<std::complex<double>>> &matrix_data)
+void read_jsonl_data(const std::string &file, ComplexMatrix &matrix_data)
 {
     std::ifstream infile(file);
     if (!infile.is_open())
@@ -50,39 +177,32 @@ void read_jsonl_data(const std::string &file,
     json j_all;
     infile >> j_all;
 
-    // 读取二维数组
     if (!j_all.contains("data") || !j_all["data"].is_array())
     {
         throw std::runtime_error("Invalid JSON: 'data' field missing or not an array");
     }
 
     const auto &data_array = j_all["data"];
-    size_t num_rows = data_array.size();
+    const size_t num_rows = data_array.size();
     if (num_rows == 0)
-        return;
-
-    size_t num_cols = data_array[0].size();
-
-    for (size_t row = 0; row < num_rows; ++row)
     {
-        const auto &row_data = data_array[row];
-        if (!row_data.is_array() || row_data.size() != num_cols)
-        {
-            throw std::runtime_error("Inconsistent row size in JSON data");
-        }
+        return;
+    }
 
-        // 每行复制100次
-        for (int copy = 0; copy < 100; ++copy)
-        {
-            // 计算复制后的目标行索引 = 原始行索引 * 100 + 复制次数
-            size_t target_row = row + copy * 100;
+    if (num_rows != static_cast<size_t>(kDataNums))
+    {
+        throw std::runtime_error("Invalid JSON: row count does not match expected data count");
+    }
 
-            for (size_t col = 0; col < num_cols; ++col)
-            {
-                matrix_data[col][target_row] =
-                    std::complex<double>(row_data[col].get<double>() / 40, 0.0);
-            }
-        }
+    const size_t num_cols = data_array[0].size();
+    if (num_cols != static_cast<size_t>(kDimension))
+    {
+        throw std::runtime_error("Invalid JSON: column count does not match expected dimension");
+    }
+
+    if (matrix_data.size() < num_cols * 2)
+    {
+        throw std::runtime_error("matrix_data does not have enough columns for both layouts");
     }
 
     for (size_t row = 0; row < num_rows; ++row)
@@ -93,23 +213,27 @@ void read_jsonl_data(const std::string &file,
             throw std::runtime_error("Inconsistent row size in JSON data");
         }
 
-        // 每行复制100次
-        for (int copy = 0; copy < 100; ++copy)
+        ComplexVec scaled_row(num_cols);
+        for (size_t col = 0; col < num_cols; ++col)
         {
-            // 计算复制后的目标行索引 = 原始行索引 * 100 + 复制次数
-            size_t target_row = row * 100 + copy;
+            scaled_row[col] = Complex(row_data[col].get<double>() / kInputScaleDivisor, 0.0);
+        }
 
+        for (int copy = 0; copy < kDataNums; ++copy)
+        {
+            const size_t interleaved_row = row + copy * kDataNums;
+            const size_t blocked_row = row * kDataNums + copy;
             for (size_t col = 0; col < num_cols; ++col)
             {
-                matrix_data[col + 10][target_row] =
-                    std::complex<double>(row_data[col].get<double>() / 40, 0.0);
+                matrix_data[col][interleaved_row] = scaled_row[col];
+                matrix_data[col + kDimension][blocked_row] = scaled_row[col];
             }
         }
     }
 }
 
 Ciphertext encode_and_encrypt(const CKKSEncoder &encoder, const Encryptor &encryptor,
-                              std::vector<std::complex<double>> &message, double scale)
+                              const ComplexVec &message, double scale)
 {
     Plaintext plain;
     Ciphertext ciph;
@@ -119,11 +243,12 @@ Ciphertext encode_and_encrypt(const CKKSEncoder &encoder, const Encryptor &encry
 }
 
 std::vector<Ciphertext> encode_and_encrypt(const CKKSEncoder &encoder, const Encryptor &encryptor,
-                                           std::vector<std::vector<std::complex<double>>> &message,
+                                           const ComplexMatrix &message,
                                            double scale)
 {
     std::vector<Ciphertext> vec_ciph;
-    for (int i = 0; i < message.size(); ++i)
+    vec_ciph.reserve(message.size());
+    for (size_t i = 0; i < message.size(); ++i)
     {
         Plaintext plain;
         Ciphertext ciph;
@@ -134,134 +259,60 @@ std::vector<Ciphertext> encode_and_encrypt(const CKKSEncoder &encoder, const Enc
     return vec_ciph;
 }
 
-std::vector<std::complex<double>> decrypt_and_decode(const CKKSEncoder &encoder,
-                                                     Decryptor &decryptor, const Ciphertext &ciph)
+ComplexVec decrypt_and_decode(const CKKSEncoder &encoder, Decryptor &decryptor, const Ciphertext &ciph)
 {
     Plaintext plain;
     decryptor.decrypt(ciph, plain);
-    std::vector<std::complex<double>> message;
+    ComplexVec message;
     encoder.decode(plain, message);
     return message;
 }
 
 void encode_and_encrypt(const CKKSEncoder &encoder, const Encryptor &encryptor,
-                        std::vector<std::complex<double>> &message, double scale, Ciphertext &ciph)
+                        const ComplexVec &message, double scale, Ciphertext &ciph)
 {
     Plaintext plain;
     encoder.encode(message, scale, plain);
     encryptor.encrypt(plain, ciph);
 }
 
-std::vector<Ciphertext>
-encode_and_encrypt_mt(const CKKSEncoder &encoder, const Encryptor &encryptor,
-                      std::vector<std::vector<std::complex<double>>> &message, double scale)
+std::vector<Ciphertext> encode_and_encrypt_mt(const CKKSEncoder &encoder, const Encryptor &encryptor,
+                                              const ComplexMatrix &message,
+                                              double scale)
 {
-    ThreadPool thread_pool(num_threads);
-    std::vector<Ciphertext> vec_ciph;
-    vec_ciph.resize(message.size());
-    const int work_load = (message.size() + num_threads - 1) / num_threads;
+    std::vector<Ciphertext> vec_ciph(message.size());
 
-    std::vector<future<void>> futures;
-    for (int w = 0; w < num_threads; ++w)
+#pragma omp parallel for num_threads(kNumThreads)
+    for (int i = 0; i < static_cast<int>(message.size()); ++i)
     {
-        int start = w * work_load;
-        int end = std::min<int>(start + work_load, message.size());
-
-        if (end > start)
-        {
-            auto ret = thread_pool.enqueue(
-                [&](size_t s, size_t e)
-                {
-                    for (size_t i = s; i < e; ++i)
-                    {
-                        encode_and_encrypt(encoder, encryptor, message[i], scale, vec_ciph[i]);
-                    }
-                },
-                start, end);
-            futures.push_back(std::move(ret));
-        }
-    }
-
-    for (auto &fut : futures)
-    {
-        fut.get();
+        encode_and_encrypt(encoder, encryptor, message[i], scale, vec_ciph[i]);
     }
 
     return vec_ciph;
 }
 
-void generate_glaios_mt(KeyGenerator &kgen, vector<vector<int>> step, GaloisKeys &rot_keys)
-{
-    ThreadPool thread_pool(num_threads);
-    const int work_load = (step.size() + num_threads - 1) / num_threads;
-
-    std::vector<future<void>> futures;
-    for (int w = 0; w < num_threads; ++w)
-    {
-        int start = w * work_load;
-        int end = std::min<int>(start + work_load, step.size());
-
-        if (end > start)
-        {
-            auto ret = thread_pool.enqueue(
-                [&](size_t s, size_t e)
-                {
-                    for (size_t i = s; i < e; ++i)
-                    {
-                        kgen.create_galois_keys(step[i], rot_keys);
-                    }
-                },
-                start, end);
-            futures.push_back(std::move(ret));
-        }
-    }
-
-    for (auto &fut : futures)
-    {
-        fut.get();
-    }
-}
-
 void sub_and_square(const std::shared_ptr<EvaluatorCkksBase> &ckks_eva,
-                    std::vector<Ciphertext> &ciph_data, const std::vector<Ciphertext> &ciph_query,
-                    const poseidon::RelinKeys &relin_keys, const double scale)
+                    std::vector<Ciphertext> &ciph_data,
+                    const std::vector<Ciphertext> &ciph_query,
+                    const RelinKeys &relin_keys, double scale)
 {
-    std::vector<Ciphertext> vec_ciph;
-    vec_ciph.resize(ciph_data.size());
-
-    ThreadPool thread_pool(num_threads);
-    const int work_load = (ciph_query.size() + num_threads - 1) / num_threads;
-
-    for (int w = 0; w < num_threads; ++w)
+#pragma omp parallel for num_threads(kNumThreads)
+    for (int i = 0; i < static_cast<int>(ciph_query.size()); ++i)
     {
-        int start = w * work_load;
-        int end = std::min<int>(start + work_load, ciph_query.size());
+        ckks_eva->sub(ciph_data[i], ciph_query[i], ciph_data[i]);
+        ckks_eva->sub(ciph_data[i + kDimension], ciph_query[i], ciph_data[i + kDimension]);
 
-        if (end > start)
-        {
-            thread_pool.enqueue(
-                [&](size_t s, size_t e)
-                {
-                    for (size_t i = s; i < e; ++i)
-                    {
-                        ckks_eva->sub(ciph_data[i], ciph_query[i], ciph_data[i]);
-                        ckks_eva->sub(ciph_data[i + 10], ciph_query[i], ciph_data[i + 10]);
+        ckks_eva->multiply_relin(ciph_data[i], ciph_data[i], ciph_data[i], relin_keys);
+        ckks_eva->rescale_dynamic(ciph_data[i], ciph_data[i], scale);
 
-                        ckks_eva->multiply_relin(ciph_data[i], ciph_data[i], ciph_data[i],
-                                                 relin_keys);
-                        ckks_eva->rescale_dynamic(ciph_data[i], ciph_data[i], scale);
-
-                        ckks_eva->multiply_relin(ciph_data[i + 10], ciph_data[i + 10],
-                                                 ciph_data[i + 10], relin_keys);
-                        ckks_eva->rescale_dynamic(ciph_data[i + 10], ciph_data[i + 10], scale);
-                    }
-                },
-                start, end);
-        }
+        ckks_eva->multiply_relin(ciph_data[i + kDimension], ciph_data[i + kDimension],
+                                 ciph_data[i + kDimension], relin_keys);
+        ckks_eva->rescale_dynamic(ciph_data[i + kDimension], ciph_data[i + kDimension], scale);
     }
 }
 
-void match_param_id(Ciphertext &ciph1, Ciphertext &ciph2, std::shared_ptr<EvaluatorCkksBase> eva)
+void match_param_id(Ciphertext &ciph1, Ciphertext &ciph2,
+                    std::shared_ptr<EvaluatorCkksBase> eva)
 {
     if (ciph1.level() > ciph2.level())
     {
@@ -273,32 +324,27 @@ void match_param_id(Ciphertext &ciph1, Ciphertext &ciph2, std::shared_ptr<Evalua
     }
 }
 
-void match_scale(Ciphertext &ciph1, Ciphertext &ciph2, const CKKSEncoder &encoder,
-                 std::shared_ptr<EvaluatorCkksBase> eva, double scale)
+void match_scale(Ciphertext &ciph1, Ciphertext &ciph2,
+                 const CKKSEncoder &encoder,
+                 std::shared_ptr<EvaluatorCkksBase> eva,
+                 double scale)
 {
     if (!util::are_approximate(ciph1.scale(), ciph2.scale()))
     {
         ciph1.scale() = ciph2.scale();
-        std::vector<std::complex<double>> vec_tmp(NUM, {1.0, 0.0});
+        ComplexVec vec_tmp(1, {1.0, 0.0});
         Plaintext plt_tmp;
 
-        // for ciph_1
-        {
-            encoder.encode(vec_tmp, ciph2.parms_id(), scale * scale / ciph2.scale(), plt_tmp);
-            eva->multiply_plain(ciph2, plt_tmp, ciph2);
-            eva->rescale(ciph2, ciph2);
-        }
+        encoder.encode(vec_tmp, ciph2.parms_id(), scale * scale / ciph2.scale(), plt_tmp);
+        eva->multiply_plain(ciph2, plt_tmp, ciph2);
+        eva->rescale(ciph2, ciph2);
 
-        // for ciph_2
-        {
-            encoder.encode(vec_tmp, ciph1.parms_id(), scale * scale / ciph1.scale(), plt_tmp);
-            eva->multiply_plain(ciph1, plt_tmp, ciph1);
-            eva->rescale(ciph1, ciph1);
-        }
+        encoder.encode(vec_tmp, ciph1.parms_id(), scale * scale / ciph1.scale(), plt_tmp);
+        eva->multiply_plain(ciph1, plt_tmp, ciph1);
+        eva->rescale(ciph1, ciph1);
     }
 }
 
-// sign_1 第一次sign 近似拟合
 Ciphertext sign_1(const Ciphertext &ciph, const PolynomialVector &polys_1,
                   const PolynomialVector &polys_2, const CKKSEncoder &encoder,
                   std::shared_ptr<EvaluatorCkksBase> eva, const RelinKeys &relin_keys)
@@ -312,8 +358,8 @@ Ciphertext sign_1(const Ciphertext &ciph, const PolynomialVector &polys_1,
     return ciph_result;
 }
 
-// sign_2 第二次拟合
-Ciphertext sign_2(const Ciphertext &ciph, const PolynomialVector &polys, const CKKSEncoder &encoder,
+Ciphertext sign_2(const Ciphertext &ciph, const PolynomialVector &polys,
+                  const CKKSEncoder &encoder,
                   std::shared_ptr<EvaluatorCkksBase> eva, const RelinKeys &relin_keys)
 {
     auto ciph_result = ciph;
@@ -322,11 +368,15 @@ Ciphertext sign_2(const Ciphertext &ciph, const PolynomialVector &polys, const C
     return ciph_result;
 }
 
-void writePredictions(const std::vector<int> &data, std::string predictions_file)
+void writePredictions(const std::vector<int> &data, const std::string &predictions_file)
 {
     try
     {
-        std::filesystem::create_directories(std::filesystem::path(predictions_file).parent_path());
+        const auto parent = std::filesystem::path(predictions_file).parent_path();
+        if (!parent.empty())
+        {
+            std::filesystem::create_directories(parent);
+        }
     }
     catch (const std::filesystem::filesystem_error &e)
     {
@@ -334,41 +384,29 @@ void writePredictions(const std::vector<int> &data, std::string predictions_file
         return;
     }
 
-    // 打开文件（使用trunc模式确保覆盖原有内容）
-    std::ofstream out_file(predictions_file, std::ios::trunc);
+    std::ofstream out_file(predictions_file, std::ios::app);
     if (!out_file.is_open())
     {
         std::cerr << "无法打开文件: " << predictions_file << std::endl;
         return;
     }
 
-    // 写入JSON格式内容
     out_file << "{ \"answer\": [ ";
-    for (size_t i = 0; i < 10; ++i)
+    for (size_t i = 0; i < kPredictionCount; ++i)
     {
-        if (i < data.size())
-        {
-            out_file << data[i];
-        }
-        else
-        {
-            out_file << 100;
-        }
-
-        if (i != 9)
+        out_file << (i < data.size() ? data[i] : 100);
+        if (i + 1 != kPredictionCount)
         {
             out_file << ", ";
         }
     }
-    out_file << " ] }";
-
-    // 关闭文件
-    out_file.close();
+    out_file << " ] }" << std::endl;
 }
 
 Ciphertext accumulate_top_n_block(const Ciphertext &ciph, int n, const CKKSEncoder &encoder,
-                                  const Encryptor &enc, std::shared_ptr<EvaluatorCkksBase> ckks_eva,
-                                  const GaloisKeys rot_keys)
+                                  const Encryptor &enc,
+                                  std::shared_ptr<EvaluatorCkksBase> ckks_eva,
+                                  const GaloisKeys &rot_keys)
 {
     if (n <= 0)
     {
@@ -376,29 +414,25 @@ Ciphertext accumulate_top_n_block(const Ciphertext &ciph, int n, const CKKSEncod
     }
 
     Ciphertext ciph_rotate_sum = ciph;
-
-    std::vector<std::complex<double>> zero = {{0.0, 0.0}};
+    ComplexVec zero = {{0.0, 0.0}};
     Plaintext plain_zero;
     Ciphertext ciph_sum;
     encoder.encode(zero, ciph.parms_id(), ciph.scale(), plain_zero);
     enc.encrypt(plain_zero, ciph_sum);
 
-    int cnt = 0;
-    int bottom_nth = 0;
-    const int const_n = n;
     while (n > 1)
     {
         Ciphertext ciph_tmp;
-        if (n & 1 && n != 1)
+        if (n & 1)
         {
             ckks_eva->add(ciph_sum, ciph_rotate_sum, ciph_sum);
-            ckks_eva->rotate(ciph_rotate_sum, ciph_rotate_sum, 100, rot_keys);
-            n--;
+            ckks_eva->rotate(ciph_rotate_sum, ciph_rotate_sum, kDataNums, rot_keys);
+            --n;
         }
-        n = n >> 1;
+        n >>= 1;
         if (n)
         {
-            ckks_eva->rotate(ciph_rotate_sum, ciph_tmp, n * 100, rot_keys);
+            ckks_eva->rotate(ciph_rotate_sum, ciph_tmp, n * kDataNums, rot_keys);
             ckks_eva->add(ciph_rotate_sum, ciph_tmp, ciph_rotate_sum);
         }
     }
@@ -406,21 +440,120 @@ Ciphertext accumulate_top_n_block(const Ciphertext &ciph, int n, const CKKSEncod
     return ciph_sum;
 }
 
-std::string get_current_path()
+int run_knn(const RuntimeOptions &options)
 {
-    char path_buffer[PATH_MAX] = {0};
-    // 读取当前进程的可执行文件路径（/proc/self/exe 是 Linux 特有的符号链接）
-    ssize_t len = readlink("/proc/self/exe", path_buffer, PATH_MAX - 1);
-    std::string path = std::string(path_buffer, len);
+    util::Timestacs timer;
+    util::Timestacs timer_init;
+    util::Timestacs timer_calculate;
+    timer.start();
+    timer_init.start();
 
-    // 从完整路径中截取目录部分
-    size_t pos = path.find_last_of("/");
-    if (pos != std::string::npos)
+    PoseidonFactory::get_instance()->set_device_type(DEVICE_SOFTWARE);
+    uint32_t q_def = kScaleBits;
+
+    ParametersLiteral ckks_param_literal{CKKS, options.log_degree, options.log_degree - 1, q_def, 5,
+                                         1, 0, {}, {}};
+    std::vector<uint32_t> logQTmp(18, 32);
+    std::vector<uint32_t> logPTmp(1, 60);
+    ckks_param_literal.set_log_modulus(logQTmp, logPTmp);
+    auto context = PoseidonFactory::get_instance()->create_poseidon_context(ckks_param_literal);
+    double scale = std::pow(2.0, q_def);
+
+    PublicKey public_key;
+    RelinKeys relin_keys;
+    GaloisKeys rot_keys;
+    CKKSEncoder ckks_encoder(context);
+
+    KeyGenerator kgen(context);
+    kgen.create_public_key(public_key);
+    kgen.create_relin_keys(relin_keys);
+    kgen.create_galois_keys(kRotationSteps, rot_keys);
+
+    Encryptor enc(context, public_key, kgen.secret_key());
+    Decryptor dec(context, kgen.secret_key());
+    std::shared_ptr<EvaluatorCkksBase> ckks_eva =
+        PoseidonFactory::get_instance()->create_ckks_evaluator(context);
+
+    const size_t slot_count = context.parameters_literal()->degree() >> 1;
+    std::vector<std::vector<int>> slots_index(1, std::vector<int>(slot_count, 0));
+    std::vector<int> idxF(slot_count);
+    for (size_t i = 0; i < slot_count; ++i)
     {
-        path = path.substr(0, pos + 1);  // 保留路径分隔符
+        idxF[i] = static_cast<int>(i);
     }
-    return path;
+    slots_index[0] = idxF;
+
+    const PolynomialVector polys = build_polynomial_vector(
+        {0, 3.816912, 0, -9.226954, 0, 11.954844, 0, -5.516258}, slots_index, 7);
+    const PolynomialVector polys_1 = build_polynomial_vector(
+        {0, 2.5390678487943066, 0, -15.36649590685934, 0, 72.05487340640471,
+         0, -229.83084441307128, 0, 510.7603223522984, 0, -810.2812835443851,
+         0, 932.3382320828513, 0, -783.7465043857175, 0, 480.4851545467111,
+         0, -212.16308093582333, 0, 65.63925462800184, 0, -13.490628831305791,
+         0, 1.6532569365778251, 0, -0.091371472313767},
+        slots_index, 27);
+    const PolynomialVector polys_2 =
+        build_polynomial_vector({0.5, 0.197, 0, -0.004}, slots_index, 3);
+
+    ComplexMatrix query(kDimension, ComplexVec(slot_count, {0.0, 0.0}));
+    ComplexMatrix data(kDimension * 2, ComplexVec(slot_count, {0.0, 0.0}));
+
+    read_jsonl_query(options.dataset_file, query, slot_count);
+    read_jsonl_data(options.dataset_file, data);
+
+    timer_init.end();
+    timer_calculate.start();
+
+    std::vector<Ciphertext> ciph_query = encode_and_encrypt_mt(ckks_encoder, enc, query, scale);
+    std::vector<Ciphertext> ciph_data = encode_and_encrypt_mt(ckks_encoder, enc, data, scale);
+
+    ComplexVec cmp_top_k(slot_count, {0.0, 0.0});
+    for (int i = 0; i < kDataNums; ++i)
+    {
+        cmp_top_k[i].real(kComparisonThreshold);
+    }
+    Ciphertext ciph_top_k = encode_and_encrypt(ckks_encoder, enc, cmp_top_k, scale);
+
+    sub_and_square(ckks_eva, ciph_data, ciph_query, relin_keys, scale);
+    Ciphertext ciph_distance_1 = sum_distance_columns(ckks_eva, ciph_data, 0);
+    Ciphertext ciph_distance_2 = sum_distance_columns(ckks_eva, ciph_data, kDimension);
+
+    Ciphertext ciph_result;
+    ckks_eva->sub_dynamic(ciph_distance_1, ciph_distance_2, ciph_result, ckks_encoder);
+
+    Ciphertext ciph_tmp = sign_1(ciph_result, polys, polys_1, ckks_encoder, ckks_eva, relin_keys);
+    ciph_result =
+        accumulate_top_n_block(ciph_tmp, kTopKThreshold, ckks_encoder, enc, ckks_eva, rot_keys);
+
+    match_param_id(ciph_result, ciph_top_k, ckks_eva);
+    match_scale(ciph_result, ciph_top_k, ckks_encoder, ckks_eva, scale);
+
+    ckks_eva->sub_dynamic(ciph_top_k, ciph_result, ciph_result, ckks_encoder);
+    ckks_eva->multiply_const(ciph_result, kFinalScaleFactor, scale, ciph_result, ckks_encoder);
+    ckks_eva->rescale_dynamic(ciph_result, ciph_result, scale);
+
+    Plaintext pl_margin_mask;
+    std::vector<double> margin_mask = generate_margin_mask(kDataNums);
+    ComplexVec vec_margin_mask(slot_count, {1.0, 0.0});
+    for (int i = 0; i < kDataNums; ++i)
+    {
+        vec_margin_mask[i] = margin_mask[i];
+    }
+    ckks_encoder.encode(vec_margin_mask, ciph_result.parms_id(), scale, pl_margin_mask);
+    ckks_eva->multiply_plain(ciph_result, pl_margin_mask, ciph_result);
+    ckks_eva->rescale_dynamic(ciph_result, ciph_result, scale);
+
+    ciph_result = sign_2(ciph_result, polys_2, ckks_encoder, ckks_eva, relin_keys);
+
+    const auto result_index = decrypt_and_decode(ckks_encoder, dec, ciph_result);
+    writePredictions(decode_predictions(result_index), options.predictions_file);
+
+    timer_calculate.end();
+    timer.end();
+    timer_init.print_time_ms("Init time: ");
+    timer_calculate.print_time_ms("Calculate time: ");
+    timer.print_time_ms("All time: ");
+    return 0;
 }
 
-}
-
+} // namespace KNN
